Refactor CSS styles and update page content
- Removed box shadow from article and .info-content classes
- Added transition effect to main element for smoother filter animation
- Added hover animation to .clients-item:first-child img
- Removed webkit scrollbar display for better user experience

These changes aim to improve the visual presentation and user experience
on the website. The removal of box shadow simplifies the design, while
the addition of transitions and hover animations adds interactive
elements. The removal of the webkit scrollbar display improves the
overall appearance. These changes have been made to enhance the
aesthetics and usability of the website.

Issue #123